Do not fail if grep does not return dependencies
i.e. for Elements itself

+++ setup-dependencies/install-dependencies.sh
@@ -18,13 +18,18 @@ elif [ "$ID" == "centos" ] && [ "$VERSION_ID" -lt 8 ]; then
 fi
 
 # From the CMakeLists.txt, retrieve the list of dependencies
-cmake_deps=$(grep -oP '^elements_project\(\S+\s+\S+ USE \K(\S+ (\d\.?)+)*' CMakeLists.txt)
-rpm_dev_deps=$(echo ${cmake_deps} | awk '{for(i=1;i<NF;i+=2){print $i "-devel-" $(i+1)}}')
-rpm_doc_deps=$(echo ${cmake_deps} | awk '{for(i=1;i<NF;i+=2){print $i "-doc-" $(i+1)}}')
-yum install -y ${rpm_dev_deps} ${rpm_doc_deps}
+cmake_deps=$(grep -oP '^elements_project\(\S+\s+\S+ USE \K(\S+ (\d\.?)+)*' CMakeLists.txt || true)
+
+if [ -n "$cmake_deps" ]; then
+  rpm_dev_deps=$(echo ${cmake_deps} | awk '{for(i=1;i<NF;i+=2){print $i "-devel-" $(i+1)}}')
+  rpm_doc_deps=$(echo ${cmake_deps} | awk '{for(i=1;i<NF;i+=2){print $i "-doc-" $(i+1)}}')
+  yum install -y ${rpm_dev_deps} ${rpm_doc_deps}
+fi
 
 # Common dependencies
 yum install -y cmake make gcc-c++ rpm-build gettext
 
 # Install dependency list
-envsubst < "$1" | xargs yum install -y
+if [ -f "$1" ]; then
+  envsubst < "$1" | xargs yum install -y
+fi
